refactor: move profiler out of src/ to root memory_profiler (fixes #680)
