+++ Makefile
@@ -23,7 +23,7 @@ CFLAGS = -Wall -Wextra -Werror -std=c11 -g \
 	-Isrc/string_algorithms \
 	-Isrc/backtracking \
 	-Isrc/process_synchronization \
-	-Isrc/profiler \
+	-Imemory_profiler \
 	-Isrc/debugger \
 	-Ibenchmark
 	# -Itui
@@ -48,7 +48,7 @@ SRC_DIRS = \
 	src/string_algorithms \
 	src/backtracking \
 	src/process_synchronization \
-	src/profiler \
+	memory_profiler \
 	src/debugger \
 	benchmark
 
@@ -203,7 +203,7 @@ ifneq ($(wildcard tests/debugger/test_step_debugger.c),)
 TEST_BINS += test_step_debugger
 endif
 
-ifneq ($(wildcard tests/profiler/test_memory_tracker.c),)
+ifneq ($(wildcard tests/memory_profiler/test_memory_tracker.c),)
 TEST_BINS += test_memory_tracker
 endif
 
@@ -784,7 +784,7 @@ $(TEST_DIR)/test_step_debugger$(EXE): $(OBJS) tests/debugger/test_step_debugger.
 test_memory_tracker: $(TEST_DIR)/test_memory_tracker$(EXE)
 	$(TEST_DIR)/test_memory_tracker$(EXE)
 
-$(TEST_DIR)/test_memory_tracker$(EXE): $(OBJS) tests/profiler/test_memory_tracker.c
+$(TEST_DIR)/test_memory_tracker$(EXE): $(OBJS) tests/memory_profiler/test_memory_tracker.c
 	@$(call MKDIR_P,$(TEST_DIR))
 	$(CC) $(CFLAGS) $^ -o $@ $(LDFLAGS)
